+++ CHANGELOG.md
@@ -6,6 +6,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Added
+- `SLink` tests.
+- `SBadge` tests.
+- `SAvatar` tests.
+- `SAccordion` tests.
+- Testing environment.
 
 ## 2.0.0-beta.82 - 2024-04-04
 ### Added