+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+
+## 2.0.0-beta.83 - 2024-04-10
 ### Fixed
 - Error on `SFilter` when try to focus first element on popover.
 - Warning for types on `SFilter` use options interface.